change the architecture of the project, switch Dagger to the Hilt
