--- app/src/main/java/com/revakovskyi/eventifycommunity/di/AppComponent.kt
@@ -1,20 +1,0 @@
-package com.revakovskyi.eventifycommunity.di
-
-import android.app.Application
-import dagger.BindsInstance
-import dagger.Component
-import javax.inject.Singleton
-
-@Singleton
-@Component(modules = [AppModule::class])
-interface AppComponent {
-
-    @Component.Builder
-    interface Builder {
-
-        @BindsInstance
-        fun application(application: Application): Builder
-        fun create(): AppComponent
-    }
-
-}
