--- authorization/src/main/java/com/revakovskyi/authorization/navigation/AuthNavigationRoute.kt
@@ -1,5 +1,0 @@
-package com.revakovskyi.authorization.navigation
-
-import com.revakovskyi.core.navigation.NavigationRoute
-
-interface AuthNavigationRoute : NavigationRoute
