--- authorization/src/main/java/com/revakovskyi/authorization/navigation/Screens.kt
@@ -1,7 +1,0 @@
-package com.revakovskyi.authorization.navigation
-
-internal sealed class Screens(val route: String) {
-
-    object AuthScreen : Screens("auth_screen")
-
-}
