--- authorization/src/main/java/com/revakovskyi/authorization/presentation/AuthScreen.kt
@@ -1,45 +1,0 @@
-package com.revakovskyi.authorization.presentation
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.material3.Button
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.navigation.NavController
-import androidx.navigation.compose.rememberNavController
-import com.revakovskyi.core.navigation.MainRoutes
-
-@Composable
-fun AuthScreen(
-    navController: NavController,
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier.fillMaxSize(),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text(text = "Auth screen")
-
-        Spacer(modifier = modifier.height(50.dp))
-
-        Button(
-            onClick = { navController.navigate(MainRoutes.PROFILE_ROUT) },
-        ) {
-            Text(text = "To Profile Screen")
-        }
-    }
-}
-
-@Composable
-@Preview(showSystemUi = true, showBackground = true)
-fun ShowAuthScreen() {
-    AuthScreen(navController = rememberNavController())
-}
