+++ domain/build.gradle.kts
@@ -1,0 +1,16 @@
+plugins {
+    id("kotlin")
+    alias(libs.plugins.kotlin.jvm)
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_1_8
+    targetCompatibility = JavaVersion.VERSION_1_8
+}
+
+dependencies {
+
+    implementation(libs.hilt.inject)
+    implementation(libs.coroutines)
+
+}
