+++ featureAuth/src/main/AndroidManifest.xml
