+++ featureAuth/src/main/java/com/revakovskyi/featureauth/navigation/AuthNavigationRoute.kt
@@ -1,0 +1,5 @@
+package com.revakovskyi.featureauth.navigation
+
+import com.revakovskyi.core.navigation.NavigationRoute
+
+interface AuthNavigationRoute : NavigationRoute
