+++ featureAuth/src/main/java/com/revakovskyi/featureauth/navigation/Screens.kt
@@ -1,0 +1,7 @@
+package com.revakovskyi.featureauth.navigation
+
+internal sealed class Screens(val route: String) {
+
+    object AuthScreen : Screens("auth_screen")
+
+}
