+++ featureProfile/src/main/AndroidManifest.xml
