+++ featureProfile/src/main/java/com/revakovskyi/featureprofile/navigation/ProfileNavigationRoute.kt
@@ -1,0 +1,5 @@
+package com.revakovskyi.featureprofile.navigation
+
+import com.revakovskyi.core.navigation.NavigationRoute
+
+interface ProfileNavigationRoute : NavigationRoute
