+++ featureProfile/src/main/java/com/revakovskyi/featureprofile/navigation/Screens.kt
@@ -1,0 +1,7 @@
+package com.revakovskyi.featureprofile.navigation
+
+internal sealed class Screens(val route: String) {
+
+    object ProfileScreen : Screens("profile_screen")
+
+}
