+++ featureSplash/src/main/java/com/revakovskyi/featuresplash/navigation/Screens.kt
@@ -1,0 +1,7 @@
+package com.revakovskyi.featuresplash.navigation
+
+internal sealed class Screens(val route: String) {
+
+    object SplashScreen : Screens("splash_screen")
+
+}
