+++ featureSplash/src/main/java/com/revakovskyi/featuresplash/navigation/SplashNavigationRoute.kt
@@ -1,0 +1,5 @@
+package com.revakovskyi.featuresplash.navigation
+
+import com.revakovskyi.core.navigation.NavigationRoute
+
+interface SplashNavigationRoute : NavigationRoute
