+++ localDb/build.gradle.kts
@@ -50,10 +50,12 @@ android {
 
 dependencies {
 
+    // Android
     implementation(libs.android.coreKtx)
     implementation(libs.android.appcompat)
     implementation(libs.android.material)
 
+    // Room
     implementation(libs.bundles.room)
     annotationProcessor(libs.room.compiler)
     ksp(libs.room.compiler)
