--- profile/src/main/java/com/revakovskyi/profile/navigation/ProfileNavigationRoute.kt
@@ -1,5 +1,0 @@
-package com.revakovskyi.profile.navigation
-
-import com.revakovskyi.core.navigation.NavigationRoute
-
-interface ProfileNavigationRoute : NavigationRoute
