--- profile/src/main/java/com/revakovskyi/profile/navigation/Screens.kt
@@ -1,7 +1,0 @@
-package com.revakovskyi.profile.navigation
-
-internal sealed class Screens(val route: String) {
-
-    object ProfileScreen : Screens("profile_screen")
-
-}
