--- splashScreen/src/main/java/com/revakovskyi/splashscreen/navigation/Screens.kt
@@ -1,7 +1,0 @@
-package com.revakovskyi.splashscreen.navigation
-
-internal sealed class Screens(val route: String) {
-
-    object SplashScreen : Screens("splash_screen")
-
-}
