--- splashScreen/src/main/java/com/revakovskyi/splashscreen/navigation/SplashNavigationRoute.kt
@@ -1,5 +1,0 @@
-package com.revakovskyi.splashscreen.navigation
-
-import com.revakovskyi.core.navigation.NavigationRoute
-
-interface SplashNavigationRoute : NavigationRoute
